requirements.txt
# python version 3.11
openai==1.13.3
llama_index==0.7.20
fastapi==0.101.0
uvicorn==0.23.2
python-dotenv==1.0.0
gunicorn==22.0.0